requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
groq==0.4.1
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
PyPDF2==3.0.1
python-docx==0.8.11
beautifulsoup4==4.12.2
requests==2.31.0
plotly==5.17.0
sentence-transformers==2.2.2
transformers==4.35.0
textstat==0.7.3
python-dotenv==1.0.0